# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hijrah App API",
        "description": "Complete API documentation for Hijrah App - A comprehensive assessment and program management system",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/methodology/{methodologyId}/answers": {
            "get": {
                "tags": [
                    "User Answers"
                ],
                "summary": "Get methodology answers",
                "description": "Retrieve user answers for a methodology assessment",
                "operationId": "getMethodologyAnswers",
                "parameters": [
                    {
                        "name": "methodologyId",
                        "in": "path",
                        "description": "Methodology ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAnswersResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Methodology not found or no answers submitted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Answers"
                ],
                "summary": "Submit methodology answers",
                "description": "Submit user answers for a methodology assessment",
                "operationId": "submitMethodologyAnswers",
                "parameters": [
                    {
                        "name": "methodologyId",
                        "in": "path",
                        "description": "Methodology ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "User answers data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitAnswersRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Answers submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubmitAnswersResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Methodology not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/methodology/{methodologyId}/pillar/{pillarId}/answers": {
            "get": {
                "tags": [
                    "User Answers"
                ],
                "summary": "Get pillar answers",
                "description": "Retrieve user answers for a pillar assessment",
                "operationId": "getPillarAnswers",
                "parameters": [
                    {
                        "name": "methodologyId",
                        "in": "path",
                        "description": "Methodology ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "pillarId",
                        "in": "path",
                        "description": "Pillar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAnswersResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Methodology, pillar not found or no answers submitted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Answers"
                ],
                "summary": "Submit pillar answers",
                "description": "Submit user answers for a pillar assessment",
                "operationId": "submitPillarAnswers",
                "parameters": [
                    {
                        "name": "methodologyId",
                        "in": "path",
                        "description": "Methodology ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "pillarId",
                        "in": "path",
                        "description": "Pillar ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "User answers data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitAnswersRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pillar answers submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubmitAnswersResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Methodology or pillar not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/methodology/{methodologyId}/module/{moduleId}/answers": {
            "get": {
                "tags": [
                    "User Answers"
                ],
                "summary": "Get module answers",
                "description": "Retrieve user answers for a module assessment",
                "operationId": "getModuleAnswers",
                "parameters": [
                    {
                        "name": "methodologyId",
                        "in": "path",
                        "description": "Methodology ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "moduleId",
                        "in": "path",
                        "description": "Module ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserAnswersResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Methodology, module not found or no answers submitted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Answers"
                ],
                "summary": "Submit module answers",
                "description": "Submit user answers for a module assessment",
                "operationId": "submitModuleAnswers",
                "parameters": [
                    {
                        "name": "methodologyId",
                        "in": "path",
                        "description": "Methodology ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "moduleId",
                        "in": "path",
                        "description": "Module ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "User answers data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitAnswersRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Module answers submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubmitAnswersResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Methodology or module not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubmitAnswersRequest": {
                "required": [
                    "answers"
                ],
                "properties": {
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserAnswerSubmission"
                        }
                    }
                },
                "type": "object"
            },
            "UserAnswerSubmission": {
                "required": [
                    "question_id",
                    "answer_id"
                ],
                "properties": {
                    "question_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "answer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "SubmitAnswersResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Answers submitted successfully"
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "example": 85.5
                    },
                    "total_questions": {
                        "type": "integer",
                        "example": 10
                    },
                    "answered_questions": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "UserAnswer": {
                "properties": {
                    "question_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "answer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "type": "integer",
                        "example": 5
                    },
                    "question_text": {
                        "type": "string",
                        "example": "How would you rate...?"
                    },
                    "answer_text": {
                        "type": "string",
                        "example": "Excellent"
                    },
                    "submitted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserAnswersResponse": {
                "properties": {
                    "methodology_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "pillar_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "module_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "example": 85.5
                    },
                    "completion_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserAnswer"
                        }
                    },
                    "submitted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a bearer token to access this endpoint",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}